requirements.txt
numpy
numba
cupy-cuda13x